fix(rpc): widen signature v to u128/U256 for full EIP-155 support
The `signature_v` function was truncating the EIP-155 `v` value from
`u128` to `u64` with a silent fallback to `u64::MAX` on overflow. For
legacy transactions with very large chain IDs, this would persist an
invalid `v` value, breaking signature verification and producing
nonsensical RPC responses.

Widen the `v` field to `u128` in `IndexedTransaction` and `TxMetadata`
(matching the `to_eip155_value` return type), and to `U256` in
`RpcTransaction` (matching Ethereum JSON-RPC convention). This
eliminates the truncation entirely so the full EIP-155 range is
represented without data loss.

+++ crates/storage/indexer/src/types.rs
@@ -54,8 +54,8 @@ pub struct IndexedTransaction {
     pub max_fee_per_gas: Option<u128>,
     /// Max priority fee per gas (EIP-1559 and later typed transactions).
     pub max_priority_fee_per_gas: Option<u128>,
-    /// V component of the transaction signature.
-    pub v: u64,
+    /// V component of the transaction signature (u128 to represent the full EIP-155 range).
+    pub v: u128,
     /// R component of the transaction signature.
     pub r: U256,
     /// S component of the transaction signature.
